+++ pip_services3_rpc/services/HttpEndpoint.py
@@ -8,7 +8,6 @@
     :copyright: Conceptual Vision Consulting LLC 2018-2019, see AUTHORS for more details.
     :license: MIT, see LICENSE for more details.
 """
-import inspect
 import json
 import time
 from threading import Thread
@@ -275,19 +274,6 @@ class HttpEndpoint(IOpenable, IConfigurable, IReferenceable):
                     correlation_id = None if not params else params.get('correlation_id')
                     schema.validate_and_throw_exception(correlation_id, params, False)
 
-                    # prepare parameters for handler, TODO: maybe need refactor
-                    if request.method == "GET":
-                        handler_arg_list = inspect.getfullargspec(handler).args
-                        if handler_arg_list[0] == 'self': del handler_arg_list[0]
-                        uniq = set(kwargs.keys()) | set(handler_arg_list)
-                        kwargs.update(params)
-                        new_kwargs = {}
-                        for k in kwargs.keys():
-                            if k in uniq and k != 'correlation_id':
-                                new_kwargs[k] = kwargs[k]
-                        kwargs = new_kwargs
-                        del new_kwargs
-
                 return handler(*args, **kwargs)
             except Exception as ex:
                 # hack the redirect response in bottle
@@ -379,6 +365,7 @@ class HttpEndpoint(IOpenable, IConfigurable, IReferenceable):
         :param authorize: the authorization interceptor
         :param action: the action to perform at the given route.
         """
+
         def action_with_authorize(*args, **kwargs):
             # hack to pass the parameters in authorizer
             bottle.request.params['kwargs'] = kwargs