+++ pip_services3_rpc/services/CommandableHttpService.py
@@ -85,8 +85,8 @@ class CommandableHttpService(RestService):
 
     def __get_handler(self, command: ICommand) -> Callable:
         def handler():
-            params = self.get_data()
-            correlation_id = params['correlation_id'] if 'correlation_id' in params else None
+            params = self._get_data()
+            correlation_id = params.get('correlation_id')
             args = Parameters.from_value(params)
             timing = self._instrument(correlation_id, self._base_route + '.' + command.get_name())
             try: