+++ CHANGELOG.md
@@ -1,10 +1,11 @@
 # <img src="https://uploads-ssl.webflow.com/5ea5d3315186cf5ec60c3ee4/5edf1c94ce4c859f2b188094_logo.svg" alt="Pip.Services Logo" width="200"> <br/> Remote Procedure Calls for Python Changelog
 
-## <a name="3.2.5"></a> 3.2.5 (2021-08-25)
+## <a name="3.2.5-3.2.6"></a> 3.2.5-3.2.6 (2021-08-30)
 
 ### Bug Fixes
 * Fixed getting correlation id from request
 * Update getting query parameters
+* Fixed preparing empty params in services
 
 ## <a name="3.2.2 - 3.2.4"></a> 3.2.2 - 3.2.4 (2021-06-17)